+++ OmniPort.UI.Presentation/Mapping/OmniPortMappingProfile.cs
@@ -2,6 +2,7 @@
 using OmniPort.Core.Models;
 using OmniPort.Core.Records;
 using OmniPort.Data;
+using OmniPort.UI.Presentation.Models;
 using System.Collections.Generic;
 using System.Linq;
