+++ OmniPort.UI.Presentation/Models/MappingTemplateForm.cs
@@ -5,16 +5,17 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace OmniPort.Core.Models
+namespace OmniPort.UI.Presentation.Models
 {
-    public class TemplateEditForm
+    public class MappingTemplateForm
     {
         public int? Id { get; set; } // null => Create
 
         [Required] public string Name { get; set; } = string.Empty;
-        [Required] public SourceType SourceType { get; set; }
 
-        [MinLength(1)]
-        public List<TemplateFieldRow> Fields { get; set; } = new();
+        [Required] public int SourceTemplateId { get; set; }
+        [Required] public int TargetTemplateId { get; set; }
+
+        public Dictionary<int, int?> TargetToSource { get; set; } = new();
     }
 }
