+++ OmniPort.UI.Presentation/Models/TemplateFieldRow.cs
@@ -1,11 +1,12 @@
-﻿using System;
+﻿using OmniPort.Core.Models;
+using System;
 using System.Collections.Generic;
 using System.ComponentModel.DataAnnotations;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace OmniPort.Core.Models
+namespace OmniPort.UI.Presentation.Models
 {
     public class TemplateFieldRow
     {
