+++ OmniPort.UI.Presentation/Models/TransformationRunForm.cs
@@ -5,7 +5,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace OmniPort.Core.Models
+namespace OmniPort.UI.Presentation.Models
 {
     public class TransformationRunForm
     {
