+++ OmniPort.UI.Presentation/ViewModels/TemplateEditorViewModel.cs
@@ -2,6 +2,7 @@
 using OmniPort.Core.Models;
 using OmniPort.Core.Records;
 using OmniPort.UI.Presentation.Interfaces;
+using OmniPort.UI.Presentation.Models;
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
